[REF][SLOT-283]: refactor existing cron jobs
